requirements.txt
fastapi == 0.115.0
uvicorn == 0.31.0
SQLAlchemy == 2.0.35
passlib[bcrypt] == 1.7.4
psycopg2-binary == 2.9.9
pyjwt == 2.9.0
python-dotenv == 1.0.1
pydantic[email] == 2.2.0
python-multipart == 0.0.12
fastapi_auth_middleware == 1.0.2
Requests == 2.32.3
fastapi-pagination == 0.12.31
python-json-logger == 2.0.7